SA-0MLXEM41U1VIK3TB: Extract delegation orchestration from scheduler.py into delegation.py
Move ~730 lines of delegation orchestration and report generation code
from scheduler.py into DelegationOrchestrator class in delegation.py:

- Module-level helpers: _summarize_for_discord, _content_hash, _trim_text,
  _format_in_progress_items, _format_candidate_line, _build_dry_run_report,
  _build_dry_run_discord_message, _build_delegation_report,
  _build_delegation_discord_message
- Scheduler methods: _inspect_idle_delegation, _is_delegation_report_changed,
  _run_delegation_report, _run_idle_delegation, _engine_rejections,
  _recover_stale_delegations -> DelegationOrchestrator methods
- Replace delegation branch in Scheduler.start_command() with single
  DelegationOrchestrator.execute() dispatch
- Kill duplicated functions in scheduler_cli.py and triage_audit.py,
  replace with imports from ampa.delegation
- Re-export moved functions from scheduler.py for backward compatibility
- Sync webhook_module and selection references in _sync_orchestrator()
  to prevent test hangs from stale module references

Net result: +806/-1233 lines across 4 files. All 532 tests pass.
